Background location first implementation
Finally, a background location has been implemented. Now the
functionality is as follows:
Whenever the user presses the slide button on the profile view a method
will be called that sends the locations periodically and whenever the
button is pressed again the background location will stop.

There was also a getLocation method implemented that takes the location
of the user.