Added a form to introduce a title and a comment
Added a form where the user can input a title and a comment that will be
added to the location so that later the user can see it in the webpage.
There was also implementad two buttons: one for sending it with those
two fields and another one to send it with a photo. The functionality is
not yet implemented.